feat: microservicio de pagos 100% operativo, integrado con gateway y canchas

+++ src/main/java/cl/municipalidad/pagos/dto/response/DtoCanchaResponse.java
@@ -4,15 +4,21 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * DTO que actúa como espejo para recibir el JSON remoto devuelto por ms-canchas.
+ * Recopila los datos necesarios de la infraestructura para el enriquecimiento del cobro.
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class DtoCanchaResponse {
 
-    private Integer idCancha;
-    private String nombre;         // Ej: "Cancha Municipal N° 1"
+    // 🔄 CORREGIDO: Cambiado de Integer a Long para alinearse con ms-canchas y solucionar la marca roja del servicio
+    private Long idCancha;
+    
+    private String nombre;           // Ej: "Cancha Principal de Pasto Sintético"
     private String tipoDeSuperficie; // Ej: "Pasto Sintético", "Maicillo", "Parquet"
-    private String ubicacion;      // Ej: "Complejo Deportivo Estadio Municipal"
-    private String comuna;         // Ej: "Santiago", "Providencia", etc.
-    private Integer valorPorHora;  // Adaptamos la capacidad por un valor numérico coherente para el cobro
+    private String ubicacion;        // Ej: "Complejo Deportivo Estadio Municipal"
+    private String comuna;           // Ej: "Pedro Aguirre Cerda"
+    private Integer valorPorHora;    // Valor numérico coherente para el cálculo del cobro
 }
