+++ src/main/java/cl/municipalidad/pagos/model/PagosModel.java
@@ -11,6 +11,10 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Entidad de persistencia que mapea la tabla 'pagos' en la base de datos distribuida 'mysql-pagos'.
+ * Guarda la trazabilidad de los arriendos e incorpora datos enriquecidos asíncronamente.
+ */
 @Entity
 @Table(name = "pagos")
 @Data
@@ -29,11 +33,12 @@ public class PagosModel {
     @Column(name = "estado_pago", nullable = false)
     private String estadoPago; // Ej: "PENDIENTE", "PAGADO", "RECHAZADO"
 
+    // 🔄 CORREGIDO: Cambiado de Integer a Long para mantener estricta consistencia con el idCancha de ms-canchas
     @Column(name = "id_cancha", nullable = false)
-    private Integer idCancha; // ID de la cancha que se está arrendando (Validado externamente)
+    private Long idCancha; 
 
     @Column(name = "nombre_cancha", nullable = false)
-    private String nombreCancha; // Nombre descriptivo obtenido del otro microservicio
+    private String nombreCancha; // Nombre descriptivo obtenido mediante WebClient desde ms-canchas
 
     @Column(name = "fecha_pago", nullable = false)
     private LocalDate fechaPago;
